fix: extract CustomDrawer from header to drawer
