+++ frontend/src/components/header.tsx
@@ -1,26 +1,17 @@
-import React, { useState } from 'react';
+import React, { useState, useCallback } from 'react';
 
 import styled from '@emotion/styled';
-import MailIcon from '@mui/icons-material/Mail';
 import MenuIcon from '@mui/icons-material/Menu';
-import InboxIcon from '@mui/icons-material/MoveToInbox';
 import SearchIcon from '@mui/icons-material/Search';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
-import Divider from '@mui/material/Divider';
-import Drawer from '@mui/material/Drawer';
-import MuiDrawer from '@mui/material/Drawer';
 import IconButton from '@mui/material/IconButton';
 import InputBase from '@mui/material/InputBase';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import ListItemIcon from '@mui/material/ListItemIcon';
-import ListItemText from '@mui/material/ListItemText';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 
-const drawerWidth = 240;
+import CustomDrawer from './drawer';
 
 const StyledAppBar = styled(AppBar)({
   position: 'fixed',
@@ -52,31 +43,12 @@ const SearchInputBase = styled(InputBase)({
   zIndex: 2,
 });
 
-interface StyledDrawerProps {
-  drawerOpened : boolean,
-}
-
-const StyledDrawer = styled(MuiDrawer)({
-  flexShrink: 0,
-  whiteSpace: 'nowrap',
-  boxSizing: 'border-box',
-  anchor: 'left',
-}, ({ drawerOpened }: StyledDrawerProps)=> ({
-  ...(drawerOpened && {
-    width: 60,
-    '& .MuiDrawer-paper': { width: 60 },
-  }),
-  ...(!drawerOpened && {
-    width: 120,
-    '& .MuiDrawer-paper': 120,
-  }),
-}));
 export default function Header() {
   const [opened, setOpen] = useState(true);
 
-  const handleDrawer = () => {
+  const handleDrawer = useCallback(() => {
     setOpen(!opened);
-  };
+  }, [setOpen, opened]);
 
   return (
     <Box>
@@ -99,30 +71,7 @@ export default function Header() {
         </Button>
       </StyledToolbar>
     </StyledAppBar>
-    <StyledDrawer variant="permanent" drawerOpened={opened}>
-    <Toolbar />
-    <List>
-      {['홈', '탐색', '구독', 'originals'].map((text, index) => (
-        <ListItem button key={text}>
-          <ListItemIcon>
-            {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-          </ListItemIcon>
-          <ListItemText primary={text} />
-        </ListItem>
-      ))}
-    </List>
-    <Divider />
-    <List>
-      {['홈', '탐색', '구독', 'originals'].map((text, index) => (
-        <ListItem button key={text}>
-          <ListItemIcon>
-            {index % 2 === 0 ? <InboxIcon /> : <MailIcon />}
-          </ListItemIcon>
-          <ListItemText primary={text} />
-        </ListItem>
-      ))}
-    </List>
-  </StyledDrawer>
+    <CustomDrawer opened={opened}/>
   </Box>
   );
 }
